Controller and Factory
Added a try-catch to Controller and Factory. I'm wondering why we have to make the error message in AddPollView, because the User is never able to change the number of Polls and so will never add a poll beyond the PollList Length. Therefore the error message will never show. Should this instead be the error that pops up once the user has already changed all of the options?
Same situation in the Factory, the catch will never be used so there's no handling right now.